HEpiR-HREvolution/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
python-multipart==0.0.12
pypdf==5.1.0
python-docx==1.1.2
